molgenis-api-permissions/src/main/java/org/molgenis/api/permissions/exceptions/ReadPermissionDeniedException.java
package org.molgenis.api.permissions.exceptions;

import static java.util.Objects.requireNonNull;

import org.molgenis.i18n.CodedRuntimeException;

public class ReadPermissionDeniedException extends CodedRuntimeException {
  private static final String ERROR_CODE = "PRM07";

  private final String typeId;

  public ReadPermissionDeniedException(String typeId) {
    super(ERROR_CODE);
    this.typeId = requireNonNull(typeId);
  }

  @Override
  public String getMessage() {
    return String.format("typeId:%s", typeId);
  }

  @Override
  protected Object[] getLocalizedMessageArguments() {
    return new Object[] {typeId};
  }
}
